Rework radiobutton/select switching, such that the radiobutton set is the real field, and select just proxies to the radiobuttons
Previously there were handlers for both.
This changes it so only the radiobutton set is the real field.
Select is just a visual equivalent that proxies to the real radiobutton set field.
This simplifies the handlers down.
This also makes the proxying merely a local concern where the select is created; everything else happens naturally.